@@ -3,10 +3,12 @@
   "version": "1.0.6",
   "description": "i18n .po to .json file converter",
   "main": "dist/index.js",
+  "type": "module",
   "scripts": {
-    "test": "karma start --browsers Firefox --single-run",
-    "testwatch": "karma start --browsers Firefox",
-    "build": "tsc index.ts --outDir dist"
+    "lint": "eslint . --ext .js,.ts,.tsx",
+    "test": "jest",
+    "typecheck": "tsc -p . --noEmit",
+    "build": "tsc -p . --outDir dist"
   },
   "keywords": [
     "i18n"
@@ -20,29 +22,32 @@
   },
   "author": "",
   "license": "MIT",
-  "devDependencies": {
-    "@types/get-stdin": "^5.0.1",
-    "@types/yargs": "^8.0.2",
-    "array-xor": "^0.1.1",
-    "browserify": "^14.4.0",
-    "eslint": "3.13.1",
-    "i18n-proto": "1.0.5",
-    "karma": "1.7.0",
-    "karma-browserify": "^5.1.1",
-    "karma-firefox-launcher": "^1.0.1",
-    "karma-mocha": "^1.3.0",
-    "karma-typescript": "^3.0.4",
-    "mocha": "^3.4.2",
-    "ts-node": "^3.3.0",
-    "tslib": "1.5.0",
-    "tslint": "4.3.1",
-    "tslint-eslint-rules": "3.2.3",
-    "typescript": "^2.4.2",
-    "uglify-js": "3.0.18",
-    "watchify": "^3.9.0"
+  "resolutions": {
+    "wrap-ansi": "7.0.0",
+    "string-width": "4.1.0",
+    "strip-ansi": "6.0.1"
   },
-  "dependencies": {
-    "get-stdin": "5.0",
-    "yargs": "8.0"
+  "devDependencies": {
+    "@types/assert": "^1.5.10",
+    "@types/cli": "^0.11.25",
+    "@types/get-stdin": "^7.0.0",
+    "@types/glob": "^8.1.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.30",
+    "@typescript-eslint/eslint-plugin": "6.21.0",
+    "@typescript-eslint/parser": "6.21.0",
+    "cli": "^1.0.1",
+    "cross-env": "^7.0.3",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-unused-imports": "3.0.0",
+    "glob": "^10.3.10",
+    "i18n-proto": "1.0.5",
+    "jest": "^29.7.0",
+    "prettier": "3.2.5",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "tslib": "2.6.2",
+    "typescript": "5.3.3"
   }
 }
