@@ -1,10 +1,8 @@
-import * as assert from 'assert';
+import assert from 'assert';
 import { PluralI18NEntry, SingleI18NEntry, TranslationMeta } from 'i18n-proto';
 import { PoOptions } from '../src/types';
 import { overridePanic, overrideWarning } from '../src/panic';
-
-import { splitInTwo, convert, parseEntry, parseHeader, _parse } from '../src/convert';
-const xor = require('array-xor');
+import { splitInTwo, parseEntry, parseHeader, _parse } from '../src/convert';
 
 let panics = [];
 let warnings = [];
@@ -23,29 +21,29 @@ describe('PO to JSON converter: positive tests', () => {
   beforeEach(() => preparePanic());
 
   it('Splits string in two parts', () => {
-    let spacedString = 'Very long string with spaces';
-    let underscoredString = 'very_long_underscored_identifier';
-    let stringWithNoSeparator = 'stringwithoutseparator';
-    let splittedSpaced = splitInTwo(spacedString);
-    let splittedUnderscored = splitInTwo(underscoredString, '_');
-    let splittedWithoutSeparator = splitInTwo(stringWithNoSeparator);
-    assert.equal(splittedSpaced[0], 'Very');
-    assert.equal(splittedSpaced[1], 'long string with spaces');
-    assert.equal(splittedUnderscored[0], 'very');
-    assert.equal(splittedUnderscored[1], 'long_underscored_identifier');
-    assert.equal(splittedWithoutSeparator[0], 'stringwithoutseparator');
-    assert.equal(splittedWithoutSeparator[1], '');
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const spacedString = 'Very long string with spaces';
+    const underscoredString = 'very_long_underscored_identifier';
+    const stringWithNoSeparator = 'stringwithoutseparator';
+    const splittedSpaced = splitInTwo(spacedString);
+    const splittedUnderscored = splitInTwo(underscoredString, '_');
+    const splittedWithoutSeparator = splitInTwo(stringWithNoSeparator);
+    assert.strictEqual(splittedSpaced[0], 'Very');
+    assert.strictEqual(splittedSpaced[1], 'long string with spaces');
+    assert.strictEqual(splittedUnderscored[0], 'very');
+    assert.strictEqual(splittedUnderscored[1], 'long_underscored_identifier');
+    assert.strictEqual(splittedWithoutSeparator[0], 'stringwithoutseparator');
+    assert.strictEqual(splittedWithoutSeparator[1], '');
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic simple single i18n entry', () => {
-    let entry = `
+    const entry = `
       msgid "Some test \\"quoted\\" entry"
       msgstr "Nekiy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry',
       msgStr: 'Nekiy "quoted" perevod',
       comments: [],
@@ -55,14 +53,14 @@ describe('PO to JSON converter: positive tests', () => {
       msgStrPlural: [],
     };
 
-    let actual = _parse(entry.split("\n"), false, false);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), false, false);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic multiline single i18n entry', () => {
-    let entry = `
+    const entry = `
       msgid "Some "
         "test \\"quoted\\" entry "
         "multiline"
@@ -70,7 +68,7 @@ describe('PO to JSON converter: positive tests', () => {
         "Nekiy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry multiline',
       msgStr: 'I esche Nekiy "quoted" perevod',
       comments: [],
@@ -80,14 +78,14 @@ describe('PO to JSON converter: positive tests', () => {
       msgStrPlural: [],
     };
 
-    let actual = _parse(entry.split("\n"), false, false);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), false, false);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic full single i18n entry', () => {
-    let entry = `
+    const entry = `
       #. Comment 1
       #. Comment 2
       #: occurence 1
@@ -97,7 +95,7 @@ describe('PO to JSON converter: positive tests', () => {
       msgstr "Nekiy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry',
       msgStr: 'Nekiy "quoted" perevod',
       comments: ['Comment 1', 'Comment 2'],
@@ -107,21 +105,21 @@ describe('PO to JSON converter: positive tests', () => {
       msgStrPlural: [],
     };
 
-    let actual = _parse(entry.split("\n"), true, true);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), true, true);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic simple plural i18n entry', () => {
-    let entry = `
+    const entry = `
       msgid "Some test \\"quoted\\" entry"
       msgid_plural "Some plural test \\"quoted\\" entry"
       msgstr[0] "Nekiy \\"quoted\\" perevod"
       msgstr[1] "Nekiy pluralniy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry',
       msgStr: undefined,
       comments: [],
@@ -131,14 +129,14 @@ describe('PO to JSON converter: positive tests', () => {
       msgStrPlural: ['Nekiy "quoted" perevod', 'Nekiy pluralniy "quoted" perevod'],
     };
 
-    let actual = _parse(entry.split("\n"), false, false);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), false, false);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic multiline plural i18n entry', () => {
-    let entry = `
+    const entry = `
       msgid "Some "
         "test \\"quoted\\" entry "
         "multiline"
@@ -152,7 +150,7 @@ describe('PO to JSON converter: positive tests', () => {
         "pluralniy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry multiline',
       msgStr: undefined,
       comments: [],
@@ -165,14 +163,14 @@ describe('PO to JSON converter: positive tests', () => {
       ],
     };
 
-    let actual = _parse(entry.split("\n"), false, false);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), false, false);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses generic full plural i18n entry', () => {
-    let entry = `
+    const entry = `
       #. Comment 1
       #. Comment 2
       #: occurence 1
@@ -184,7 +182,7 @@ describe('PO to JSON converter: positive tests', () => {
       msgstr[1] "Nekiy pluralniy \\"quoted\\" perevod"
     `;
 
-    let expected = {
+    const expected = {
       msgid: 'Some test "quoted" entry',
       msgStr: undefined,
       comments: ['Comment 1', 'Comment 2'],
@@ -194,14 +192,14 @@ describe('PO to JSON converter: positive tests', () => {
       msgStrPlural: ['Nekiy "quoted" perevod', 'Nekiy pluralniy "quoted" perevod'],
     };
 
-    let actual = _parse(entry.split("\n"), true, true);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = _parse(entry.split("\n"), true, true);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses full PO header', () => {
-    let header = `
+    const header = `
     msgid ""
     msgstr ""
     "Project-Id-Version:  2gis-online\\n"
@@ -218,7 +216,7 @@ describe('PO to JSON converter: positive tests', () => {
     "Generated-By: Babel 2.1.1\\n"
     `;
 
-    let expected: TranslationMeta = {
+    const expected: TranslationMeta = {
       projectIdVersion: '2gis-online',
       reportMsgidBugsTo: 'online4@2gis.ru',
       potCreationDate: '2017-07-14 11:29+0700',
@@ -237,14 +235,14 @@ describe('PO to JSON converter: positive tests', () => {
     };
 
     const opts = { withMeta: 'full' } as PoOptions;
-    let actual = parseHeader(header, opts);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = parseHeader(header, opts);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Parses plurals PO header', () => {
-    let header = `
+    const header = `
     msgid ""
     msgstr ""
     "Project-Id-Version:  2gis-online\\n"
@@ -261,19 +259,19 @@ describe('PO to JSON converter: positive tests', () => {
     "Generated-By: Babel 2.1.1\\n"
     `;
 
-    let expected = {
+    const expected = {
       pluralForms: 'nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2'
     };
 
     const opts = { withMeta: 'plural' } as PoOptions;
-    let actual = parseHeader(header, opts);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = parseHeader(header, opts);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Creates valid I18NEntry from full single entry', () => {
-    let entry = `
+    const entry = `
       #. Comment 1
       #. Comment 2
       #: occurence 1
@@ -283,7 +281,7 @@ describe('PO to JSON converter: positive tests', () => {
       msgstr "Nekiy \\"quoted\\" perevod"
     `;
 
-    let expected: SingleI18NEntry = {
+    const expected: SingleI18NEntry = {
       type: 'single',
       entry: 'Some test "quoted" entry',
       context: 'Some "quoted" context',
@@ -298,14 +296,14 @@ describe('PO to JSON converter: positive tests', () => {
       ]
     };
 
-    let actual = parseEntry(entry, true, true);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = parseEntry(entry, true, true);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Creates valid I18NEntry from full plural entry', () => {
-    let entry = `
+    const entry = `
       #. Comment 1
       #. Comment 2
       #: occurence 1
@@ -317,7 +315,7 @@ describe('PO to JSON converter: positive tests', () => {
       msgstr[1] "Nekiy pluralniy \\"quoted\\" perevod"
     `;
 
-    let expected: PluralI18NEntry = {
+    const expected: PluralI18NEntry = {
       type: 'plural',
       entry: [
         'Some test "quoted" entry',
@@ -338,10 +336,10 @@ describe('PO to JSON converter: positive tests', () => {
       ]
     };
 
-    let actual = parseEntry(entry, true, true);
-    assert.deepEqual(actual, expected);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 0);
+    const actual = parseEntry(entry, true, true);
+    assert.deepStrictEqual(actual, expected);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 0);
   });
 });
 
@@ -349,123 +347,123 @@ describe('PO to JSON converter: negative tests', () => {
   beforeEach(() => preparePanic());
 
   it('Fails to parse invalid entry msgid + msgstr[N]', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgstr[0] "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry msgid + msgid_plural + msgstr', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgid_plural "Some plural test \\"quoted\\" entry"
     msgstr "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with non-existing msgid', () => {
-    let entry = `
+    const entry = `
     msgid_plural "Some plural test \\"quoted\\" entry"
     msgstr "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with existing but empty msgid', () => {
-    let entry = `
+    const entry = `
     msgid
     msgstr "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with existing but empty msgctxt', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgctxt
     msgstr "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with existing but empty msgstr', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgstr
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with existing but empty msgid_plural', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgid_plural
     msgstr[0] "Nekiy \\"quoted\\" perevod"
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Fails to parse invalid entry with existing but empty msgstr[N]', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgid_plural "Some plural test \\"quoted\\" entry"
     msgstr[0]
     `;
-    let result = parseEntry(entry, false, false);
-    assert.equal(result, undefined);
-    assert.equal(panics.length, 1);
-    assert.equal(warnings.length, 0);
+    const result = parseEntry(entry, false, false);
+    assert.strictEqual(result, undefined);
+    assert.strictEqual(panics.length, 1);
+    assert.strictEqual(warnings.length, 0);
   });
 
   it('Echoes warning for untranslated single entry', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgstr ""
     `;
-    let result = parseEntry(entry, false, false);
-    assert.notEqual(result, undefined);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 1);
+    const result = parseEntry(entry, false, false);
+    assert.notStrictEqual(result, undefined);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 1);
   });
 
   it('Echoes warning for untranslated plural entry', () => {
-    let entry = `
+    const entry = `
     msgid "Some test \\"quoted\\" entry"
     msgid_plural "Some plural test \\"quoted\\" entry"
     msgstr[0] ""
     msgstr[1] ""
     `;
-    let result = parseEntry(entry, false, false);
-    assert.notEqual(result, undefined);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 1);
+    const result = parseEntry(entry, false, false);
+    assert.notStrictEqual(result, undefined);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 1);
   });
 
   it('Echoes warning for unknown header entry', () => {
-    let entry = `
+    const entry = `
     msgid ""
     msgstr ""
       "Unknown-Entry: value\\n"
@@ -473,9 +471,9 @@ describe('PO to JSON converter: negative tests', () => {
     `;
 
     const opts = { withMeta: 'full' } as PoOptions;
-    let result = parseHeader(entry, opts);
-    assert.notEqual(result, undefined);
-    assert.equal(panics.length, 0);
-    assert.equal(warnings.length, 1);
+    const result = parseHeader(entry, opts);
+    assert.notStrictEqual(result, undefined);
+    assert.strictEqual(panics.length, 0);
+    assert.strictEqual(warnings.length, 1);
   });
 });
